requirements.txt
absl-py==0.9.0
aniso8601==8.0.0
astor==0.8.1
attrs==19.3.0
bleach==1.5.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
enum34==1.1.10
Flask==1.1.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
html5lib==0.9999999
imageio==2.8.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras
jsonschema==3.2.0
kiwisolver==1.2.0
llvmlite==0.32.1
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.4
numba==0.49.1
numpy==1.18.2
opencv-contrib-python==3.4.2.17
pandas==1.0.3
Pillow==7.1.1
protobuf==3.11.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
